chore: move most of utils in utils.ts
