+++ src/data-structure/augmented-stack.ts
@@ -1,6 +1,6 @@
 import type { Accessor } from "solid-js"
 import type { Instance } from "src/types.ts"
-import { augment } from "../augment.ts"
+import { augment } from "../utils.ts"
 import { Stack } from "./stack.ts"
 
 /** A generic stack data structure. It augments each value before pushing it onto the stack. */
