+++ src/index.ts
@@ -5,4 +5,4 @@ export { createT, createTComponent } from "./create-t.tsx"
 export { useFrame, useLoader, useThree } from "./hooks.ts"
 export { applyProps, useProps } from "./props.ts"
 export * as S3 from "./types.ts"
-export { buildGraph } from "./utils/build-graph.ts"
+export { autodispose, buildGraph } from "./utils.ts"
