--- src/utils/build-graph.ts
@@ -1,17 +1,0 @@
-import { Material, Object3D } from "three"
-
-export interface ObjectMap {
-  nodes: Record<string, Object3D>
-  materials: Record<string, Material>
-}
-// Collects nodes and materials from a THREE.Object3D
-export function buildGraph(object: Object3D): ObjectMap {
-  const data: ObjectMap = { nodes: {}, materials: {} }
-  object.traverse((obj: any) => {
-    if (obj.name) data.nodes[obj.name] = obj
-    if (obj.material && !(obj.material.name in data.materials)) {
-      data.materials[obj.material.name] = obj.material
-    }
-  })
-  return data
-}
