--- src/utils/has-colorspace.ts
@@ -1,11 +1,0 @@
-import { type Renderer, Texture } from "three"
-
-/**
- * Returns `true` with correct TS type inference if an object has a configurable color space (since r152).
- */
-export const hasColorSpace = <
-  T extends Renderer | Texture | object,
-  P = T extends Renderer ? { outputColorSpace: string } : { colorSpace: string },
->(
-  object: T,
-): object is T & P => "colorSpace" in object || "outputColorSpace" in object
