--- src/utils/remove-element-from-array.ts
@@ -1,5 +1,0 @@
-export const removeElementFromArray = (array: any[], value: any) => {
-  const index = array.indexOf(value)
-  if (index !== -1) array.splice(index, 1)
-  return array
-}
