--- src/utils/resolve.ts
@@ -1,9 +1,0 @@
-import type { Accessor } from "solid-js"
-
-export function resolve<T>(child: Accessor<T> | T, recursive = false): T {
-  return typeof child !== "function"
-    ? child
-    : recursive
-      ? resolve((child as Accessor<T>)())
-      : (child as Accessor<T>)()
-}
